Moving API to a best standard way using API as prefix in the URI and V1 as the suffix
Now you can access the methods by:

https://{address}/api/person/v1
