+++ src/main/java/br/com/erudio/data/vo/v1/PersonVO.java
@@ -1,4 +1,8 @@
-package br.com.erudio.data.vo;
+package br.com.erudio.data.vo.v1;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 import java.io.Serializable;
 import java.util.Objects;
