+++ src/main/java/br/com/erudio/services/PersonServices.java
@@ -2,7 +2,7 @@ package br.com.erudio.services;
 
 import br.com.erudio.converter.DozerConverter;
 import br.com.erudio.data.model.Person;
-import br.com.erudio.data.vo.PersonVO;
+import br.com.erudio.data.vo.v1.PersonVO;
 import br.com.erudio.repository.PersonRepository;
 import br.com.erudio.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
